refactor: consolidate type definitions into single module
- Move all type definitions from state.rs to types.rs
- Merge types/boundary.rs into types.rs
- Remove types/ directory structure
- Update all import paths from state:: to types::
- Maintain full API compatibility
- Consolidate tests into single test module

This reduces type definition files from 3 to 1, improving code organization
and making it easier to find and maintain type definitions.
