+++ sakurs-core/src/application/strategies/sequential.rs
@@ -9,7 +9,7 @@ use crate::application::{
 };
 use crate::domain::{
     language::LanguageRules, prefix_sum::PrefixSumComputer, reduce::BoundaryReducer,
-    state::PartialState,
+    types::PartialState,
 };
 use std::sync::Arc;
