--- sakurs-core/src/domain/types/tests/mod.rs
@@ -1,4 +1,0 @@
-//! Tests for domain types
-
-#[cfg(test)]
-mod boundary_tests;
